+++ server/plugin/app/model/request/productSpec.go
@@ -9,7 +9,7 @@ type ProductSpecSearch struct{
        ProductID  *int `json:"productId" form:"productId"` 
        Name  *string `json:"name" form:"name"` 
        SortValue  *int `json:"sortValue" form:"sortValue"` 
-       MerchantID  *int `json:"merchantId" form:"merchantId"` 
+       MerchantID  *int64 `json:"merchantId" form:"merchantId"` 
     request.PageInfo
     Sort  string `json:"sort" form:"sort"`
     Order string `json:"order" form:"order"`
